Fixed type reference (integer to num, country to string), created new files w types for cardDetails.js and address.js
