+++ lib/accounts.js
@@ -1,5 +1,6 @@
 import { check, checkString } from "./helpers/checks.js";
 import { Err } from "./helpers/errors.js";
+import { Address } from "./address.js";
 
 /** @external Promise */
 
@@ -253,16 +254,7 @@ import { Err } from "./helpers/errors.js";
  * @tag Accounts
  */
 
-/**
- * @typedef Address
- * @property {string} addressLine1
- * @property {string} addressLine2
- * @property {string} city
- * @property {string} stateOrProvince - 2 characters
- * @property {string} postalCode - 5 characters
- * @property {string} country - 2 characters
- * @tag Accounts
- */
+
 
 /**
  * Standard industry codes for businesses.
@@ -294,7 +286,7 @@ import { Err } from "./helpers/errors.js";
  * @typedef Responsibility
  * @property {boolean} isController
  * @property {boolean} isOwner
- * @property {integer} ownershipPercentage - Required if `isOwner` is true
+ * @property {number} ownershipPercentage - Required if `isOwner` is true
  * @property {string} jobTitle
  * @tag Accounts
  */
