+++ lib/address.js
@@ -1,0 +1,13 @@
+
+/**
+ * @typedef Address
+ * @property {string} addressLine1
+ * @property {string} addressLine2
+ * @property {string} city
+ * @property {string} stateOrProvince - 2 characters
+ * @property {string} postalCode - 5 characters
+ * @property {string} country - 2 characters
+ * @tag Accounts
+ */
+
+ exports.unused = {};
