+++ lib/cardDetails.js
@@ -1,0 +1,9 @@
+/**
+ * @typedef CardDetails
+ * @type {object}
+ * @property {string} dynamicDescriptor -An optional override of the default card statement descriptor for a single transfer.
+ * @property {"recurring"|"unscheduled"|null} merchantInitiatedType - Enum: [recurring unscheduled] Describes how the card transaction was initiated
+ * @tag Cards
+ */
+ 
+exports.unused = {};
