+++ lib/cards.js
@@ -1,5 +1,7 @@
 import { checkString } from "./helpers/checks.js";
 import { Err } from "./helpers/errors.js";
+import { Address } from "./address.js";
+
 
 /** @external Promise */
