+++ lib/enrichedProfile.js
@@ -37,7 +37,7 @@ import { Err } from "./helpers/errors.js";
  * @property {string} city
  * @property {string} stateOrProvince
  * @property {string} postalCode
- * @property {country} country
+ * @property {string} country
  * 
  * @tag Enrichment
  */
