+++ lib/paymentMethods.js
@@ -1,4 +1,4 @@
-import './bankAccountsTypedefs.js';
+import {BankAccount} from './bankAccountsTypedefs.js';
 import { checkString } from './helpers/checks.js';
 import { Err } from "./helpers/errors.js";
