+++ lib/transfers.js
@@ -1,6 +1,18 @@
 import { randomUUID } from "crypto";
 import { check, checkString } from "./helpers/checks.js";
 import { Err } from "./helpers/errors.js";
+import { Address } from "./address.js";
+
+
+/**
+ * @typedef CardDetails
+ * @type {object}
+ * @property {string} dynamicDescriptor -An optional override of the default card statement descriptor for a single transfer.
+ * @property {"recurring"|"unscheduled"|null} merchantInitiatedType - Enum: [recurring unscheduled] Describes how the card transaction was initiated
+ * @tag Cards
+ */
+
+
 
 /**
  * High-level account information associated with a payment method.
@@ -577,3 +589,4 @@ export class Transfers {
     return result;
   }
 }
+
