+++ lib/types/address.d.ts
@@ -1,0 +1,19 @@
+export var unused: {};
+export type Address = {
+    addressLine1: string;
+    addressLine2: string;
+    city: string;
+    /**
+     * - 2 characters
+     */
+    stateOrProvince: string;
+    /**
+     * - 5 characters
+     */
+    postalCode: string;
+    /**
+     * - 2 characters
+     */
+    country: string;
+};
+//# sourceMappingURL=address.d.ts.map
